Integrate new getExecutionStuff in GraphQL handler
